+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.2] - 2022-01-20
 
 ### Added
 - Utility function for creating `Expr::Reference`
@@ -146,7 +146,8 @@ Was yanked and replaced by [0.5.0]
 - async `UserFunction` support
 
 
-[unreleased]: https://github.com/mendelt/reval/compare/v0.5.1...HEAD
+[unreleased]: https://github.com/mendelt/reval/compare/v0.5.2...HEAD
+[0.5.2]: https://github.com/mendelt/reval/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/mendelt/reval/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/mendelt/reval/compare/v0.4.4...v0.5.0
 [0.4.5]: https://github.com/mendelt/reval/compare/v0.4.4...v0.4.5